Create separate component for user options
